Move xr_flexsel to its own module; functionize bits of it.
